requirements.txt
alembic==1.10.4
aniso8601==9.0.1
async-timeout==4.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
config==0.5.1
Flask==2.3.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
html5tagger==1.3.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
multidict==6.0.4
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
tracerite==1.1.0
typing_extensions==4.5.0
urllib3==1.26.15
websockets==11.0.2
Werkzeug==2.3.0
